src/Components/API/requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.62.0
h11==0.11.0
idna==2.10
mccabe==0.6.1
motor==2.3.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pymongo==3.11.1
regex==2020.11.13
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.12.3
python-dotenv==1.0.0
paho-mqtt==1.6.1
bcrypt
flask
pyjwt==2.4.0
requests
python-decouple
pandas
Werkzeug==2.3.7
fastapi_mail
python-multipart
